+++ tests/test_llmrails.py
@@ -2,14 +2,13 @@ from typing import Optional
 
 import pytest
 
-from colangflows import RailsConfig
-from colangflows.flows.runtime import Runtime
+from colangflows.rails import LLMRails, RailsConfig
 from tests.utils import FakeLLM
 
 
-@pytest.mark.asyncio
-async def test_1():
-    config = RailsConfig.parse_obj(
+@pytest.fixture
+def rails_config():
+    return RailsConfig.parse_obj(
         {
             "models": [
                 {
@@ -45,6 +44,9 @@ async def test_1():
         }
     )
 
+
+@pytest.mark.asyncio
+async def test_1(rails_config):
     llm = FakeLLM(
         responses=[
             "  express greeting",
@@ -57,42 +59,54 @@ async def test_1():
     async def compute(context: dict, what: Optional[str] = "2 + 3"):
         return eval(what)
 
-    runtime = Runtime(config=config, llm=llm)
-    runtime.register_action("compute", compute)
+    llm_rails = LLMRails(config=rails_config, llm=llm)
+    llm_rails.runtime.register_action(compute)
 
     events = [{"type": "user_said", "content": "Hello!"}]
 
-    new_events = await runtime.generate_events(events)
+    new_events = await llm_rails.runtime.generate_events(events)
 
     assert new_events == [
-        {"action_name": "generate_user_intent", "system": True, "type": "start_action"},
+        {
+            "action_name": "generate_user_intent",
+            "is_system_action": True,
+            "type": "start_action",
+        },
         {
             "action_name": "generate_user_intent",
             "events": [{"intent": "express greeting", "type": "user_intent"}],
+            "is_system_action": True,
             "return_value": None,
             "status": "success",
-            "system": True,
             "type": "action_finished",
         },
         {"intent": "express greeting", "type": "user_intent"},
         {"intent": "express greeting", "type": "bot_intent"},
-        {"action_name": "generate_bot_message", "system": True, "type": "start_action"},
+        {
+            "action_name": "generate_bot_message",
+            "is_system_action": True,
+            "type": "start_action",
+        },
         {
             "action_name": "generate_bot_message",
             "events": [{"content": "Hello! How are you?", "type": "bot_said"}],
+            "is_system_action": True,
             "return_value": None,
             "status": "success",
-            "system": True,
             "type": "action_finished",
         },
         {"content": "Hello! How are you?", "type": "bot_said"},
-        {"action_name": "generate_next_step", "system": True, "type": "start_action"},
+        {
+            "action_name": "generate_next_step",
+            "is_system_action": True,
+            "type": "start_action",
+        },
         {
             "action_name": "generate_next_step",
             "events": None,
+            "is_system_action": True,
             "return_value": None,
             "status": "success",
-            "system": True,
             "type": "action_finished",
         },
         {"type": "listen"},
@@ -101,50 +115,94 @@ async def test_1():
     events.extend(new_events)
     events.append({"type": "user_said", "content": "2 + 3"})
 
-    new_events = await runtime.generate_events(events)
+    new_events = await llm_rails.runtime.generate_events(events)
 
     assert new_events == [
-        {"action_name": "generate_user_intent", "system": True, "type": "start_action"},
+        {
+            "action_name": "generate_user_intent",
+            "is_system_action": True,
+            "type": "start_action",
+        },
         {
             "action_name": "generate_user_intent",
             "events": [{"intent": "ask math question", "type": "user_intent"}],
+            "is_system_action": True,
             "return_value": None,
             "status": "success",
-            "system": True,
             "type": "action_finished",
         },
         {"intent": "ask math question", "type": "user_intent"},
-        {"action_name": "compute", "system": False, "type": "start_action"},
+        {"action_name": "compute", "is_system_action": False, "type": "start_action"},
         {
             "action_name": "compute",
             "events": [],
+            "is_system_action": False,
             "return_value": 5,
             "status": "success",
             "type": "action_finished",
         },
         {"intent": "provide math response", "type": "bot_intent"},
-        {"action_name": "generate_bot_message", "system": True, "type": "start_action"},
+        {
+            "action_name": "generate_bot_message",
+            "is_system_action": True,
+            "type": "start_action",
+        },
         {
             "action_name": "generate_bot_message",
             "events": [{"content": "The answer is 5", "type": "bot_said"}],
+            "is_system_action": True,
             "return_value": None,
             "status": "success",
-            "system": True,
             "type": "action_finished",
         },
         {"content": "The answer is 5", "type": "bot_said"},
         {"intent": "ask if user happy", "type": "bot_intent"},
-        {"action_name": "generate_bot_message", "system": True, "type": "start_action"},
+        {
+            "action_name": "generate_bot_message",
+            "is_system_action": True,
+            "type": "start_action",
+        },
         {
             "action_name": "generate_bot_message",
             "events": [
                 {"content": "Are you happy with the result?", "type": "bot_said"}
             ],
+            "is_system_action": True,
             "return_value": None,
             "status": "success",
-            "system": True,
             "type": "action_finished",
         },
         {"content": "Are you happy with the result?", "type": "bot_said"},
         {"type": "listen"},
     ]
+
+
+@pytest.mark.asyncio
+async def test_2(rails_config):
+    llm = FakeLLM(
+        responses=[
+            "  express greeting",
+            "  ask math question",
+            '  "The answer is 5"',
+            '  "Are you happy with the result?"',
+        ]
+    )
+
+    async def compute(what: Optional[str] = "2 + 3"):
+        return eval(what)
+
+    llm_rails = LLMRails(config=rails_config, llm=llm)
+    llm_rails.runtime.register_action(compute)
+
+    messages = [{"role": "user", "content": "Hello!"}]
+    bot_message = await llm_rails.generate_async(messages=messages)
+
+    assert bot_message == {"role": "assistant", "content": "Hello! How are you?"}
+    messages.append(bot_message)
+
+    messages.append({"role": "user", "content": "2 + 3"})
+    bot_message = await llm_rails.generate_async(messages=messages)
+    assert bot_message == {
+        "role": "assistant",
+        "content": "The answer is 5\nAre you happy with the result?",
+    }